fix: 비동기 자동매매 일시 중단

+++ src/main/java/com/tradin/module/feign/service/BinanceFeignService.java
@@ -82,7 +82,7 @@ public class BinanceFeignService {
     private int extractUsdtBalance(List<FutureAccountBalanceDto> futureAccountBalanceDtos) {
         for (FutureAccountBalanceDto dto : futureAccountBalanceDtos) {
             if (dto.getAsset().equals("USDT")) {
-                return Integer.parseInt(dto.getAvailableBalance());
+                return (int) Double.parseDouble(dto.getAvailableBalance());
             }
         }
         return 0;
